requirements.txt
fastapi==0.61.1
uvicorn>=0.11.1
databases[postgresql]>=0.2.6
SQLAlchemy>=1.3.12
python-jose
passlib[bcrypt]
python-multipart==0.0.5
boto3

# dev
pytest>=5.3.2
requests>=2.22.0
asyncpg>=0.20.0
coverage>=4.5.4
